+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add new `Capybara/VisibilityMatcher` cop. ([@aried3r][])
 * Ignore String constants by `RSpec/Describe`. ([@AlexWayfer][])
 * Add `IgnoreShardGroups` option to `RSpec/ExpectInHook` cop. ([@jonatas][])
+* Add `IgnoreSharedGroups` option to `RSpec/ExpectInHook` cop. ([@jonatas][])
 
 ## 1.38.1 (2020-02-15)